pyproject.toml
[tool.poetry]
name = "hsr-math"
version = "0.1.0"
description = "Report optimal energy generation for Honkai: Star Rail characters."
authors = ["Andrew Hebert Jr. <gingergiant.me@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
pandas = "^2.1.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
